fix: share cross-platform path helper
