+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- **POSIX-rooted paths from Linux tooling stay absolute on Windows.** Before, several path resolvers used host-specific `Path::is_absolute`, so values such as `/ci/workspace/...` from Linux-generated coverage data, config inheritance, or plugin setup-file paths could be treated as relative on Windows and joined under the project root. After, the shared cross-platform absolute-path helper lives in `fallow-types`, remains re-exported for CLI callers, and the high-risk config `extends`, plugin setup-file, and health coverage path resolvers use it consistently. (Closes [#614](https://github.com/fallow-rs/fallow/issues/614).)
+
 - **Browser extension manifest entrypoints are no longer reported as unused.** Before, WebExtension and Chrome Extension projects that declared their runtime files only in `manifest.json` could report background service workers, content scripts, popup/options page scripts, and manifest-exposed local resources as `unused-files` because there was no import edge from the app graph. After, a built-in `browser-extension` plugin activates from conservative extension-shaped manifests (`manifest_version` 2 or 3 plus extension runtime keys), discovers root and nested extension manifests, and seeds MV3 service workers, MV2 background scripts, content script JS/CSS, popup/options/devtools/side-panel HTML, and MV2/MV3 `web_accessible_resources` entries as runtime entrypoints. Ordinary PWA web manifests remain ignored. (Closes [#616](https://github.com/fallow-rs/fallow/issues/616).)
 
 - **Varlock schema plugins and Vite integration packages no longer report as unused.** Before, projects that loaded provider packages from `.env.schema` decorators or used `@varlock/vite-integration` only through Vite config needed manual `ignoreDependencies` entries because fallow had no Varlock-aware plugin. After, the built-in Varlock plugin activates from `varlock`, `@varlock/*`, or `.env.schema`, keeps schema files reachable, credits package-like `@plugin(...)` schema references, and treats `varlock` plus `@varlock/vite-integration` as tooling dependencies while leaving local paths and URL/protocol plugin references ignored. (Closes [#622](https://github.com/fallow-rs/fallow/issues/622).)
