+++ crates/config/Cargo.toml
@@ -13,6 +13,7 @@ readme = "../../README.md"
 
 [dependencies]
 dunce = { workspace = true }
+fallow-types = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml_ng = { workspace = true }
